backend/requirements.txt
# Atlas AI Command Center Backend - Python 3.11+
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
websockets==12.0
psycopg2-binary==2.9.9
google-generativeai==0.8.0
httpx==0.27.0
groq==0.4.2
tenacity==8.2.3
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Added for Admissions Module
apscheduler==3.10.4
pdfplumber==0.10.3
pillow==10.2.0
jinja2==3.1.3
